added new message option for when the force was not selected

+++ ramp-database/ramp_database/cli.py
@@ -221,6 +221,7 @@ def delete_event(config, config_event, from_disk, force):
         # read the event_name
         config_event_params = read_config(config_event)
         event_name = config_event_params['ramp']['event_name']
+        event_dir = os.path.dirname(config_event)
 
         # remove event from the database
         with session_scope(config['sqlalchemy']) as session:
@@ -234,23 +235,33 @@ def delete_event(config, config_event, from_disk, force):
                     path_to_submissions = db_event.path_ramp_submissions
                     if os.path.exists(path_to_submissions):
                         shutil.rmtree(path_to_submissions)
-                        click.echo("""removed
-                        directory {}""".format(path_to_submissions))
+                        click.echo("removed directory {}" +
+                                   "{}".format(path_to_submissions))
                     else:
-                        click.secho("""No such submissions
-                        directory: {}""".format(path_to_submissions),
+                        click.secho("No such submissions directory: " +
+                                    "{}".format(path_to_submissions),
                                     fg='red', err=True)
             else:
-                click.secho("""No such event in the
-                            database: {}""".format(event_name),
+                click.secho("No such event in the " +
+                            "database: {}".format(event_name),
                             fg='red', err=True)
+                if from_disk and not force:
+                    click.echo("please use options --force --from_disk " +
+                               "if you would like to remove from the " +
+                               "disk events directory: {}".format(event_dir))     
 
-            event_dir = os.path.dirname(config_event)
-            if (db_event or force) and from_disk:
-                # remove this event from the disk
-                shutil.rmtree(event_dir)
-                click.echo("removed directory: {}".format(event_dir))
-            else:
+            if event_dir:
+                if (db_event or force) and from_disk:
+                    # remove this event from the disk
+                    shutil.rmtree(event_dir)
+                    click.echo("removed directory: {}".format(event_dir))
+                elif db_event:
+                    # event was removed from the db,
+                    # but it still exists on the disk
+                    click.echo("please use options --force --from_disk " +
+                               "if you would also like to remove from the " +
+                               "disk events directory: {}".format(event_dir))
+            elif (db_event or force) and from_disk:
                 click.secho("No such directory: {}".format(event_dir),
                             err=True, fg='red')
